@@ -108,7 +108,7 @@ class CloudinaryApiClient extends CloudinaryApi {
       cloudinaryResponse = CloudinaryResponse.fromJsonMap(response.data);
     } catch (error, stacktrace) {
       print('Exception occurred: $error stackTrace: $stacktrace');
-      if (error is DioError) statusCode = error.response?.statusCode;
+      if (error is DioException) statusCode = error.response?.statusCode;
       cloudinaryResponse = CloudinaryResponse.fromError('$error');
     }
     cloudinaryResponse.statusCode = statusCode;
@@ -184,7 +184,7 @@ class CloudinaryApiClient extends CloudinaryApi {
       cloudinaryResponse = CloudinaryResponse.fromJsonMap(response.data);
     } catch (error, stacktrace) {
       print('Exception occurred: $error stackTrace: $stacktrace');
-      if (error is DioError) statusCode = error.response?.statusCode;
+      if (error is DioException) statusCode = error.response?.statusCode;
       cloudinaryResponse = CloudinaryResponse.fromError('$error');
     }
     cloudinaryResponse.statusCode = statusCode;
@@ -238,7 +238,7 @@ class CloudinaryApiClient extends CloudinaryApi {
       cloudinaryResponse = CloudinaryResponse.fromJsonMap(response.data);
     } catch (error, stacktrace) {
       print('Exception occurred: $error stackTrace: $stacktrace');
-      if (error is DioError) statusCode = error.response?.statusCode;
+      if (error is DioException) statusCode = error.response?.statusCode;
       cloudinaryResponse = CloudinaryResponse.fromError('$error');
     }
     cloudinaryResponse.statusCode = statusCode;
