@@ -69,8 +69,8 @@ class Utility {
                       onTap: onImageRemove,
                       child: Container(
                         padding: const EdgeInsets.all(16),
-                        child: Row(
-                          children: const <Widget>[
+                        child: const Row(
+                          children: <Widget>[
                             Icon(
                               Icons.delete,
                               color: Colors.red,
